analytics_app/end_url/az_func_app/pyproject.toml
[tool.poetry]
name = "az-func-trial"
version = "0.1.0"
description = ""
authors = ["SanjayaElvitigala <elvitigalasanjaya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "^1.21.3"
pynacl = "^1.5.0"
cryptography = "43.0.3"
gspread = "6.1.2"
numpy = "^2.1.3"
gcloud = "^0.18.3"
pandas = "^2.2.3"
requests = "^2.32.3"
matplotlib = "^3.9.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
